wmfd-2o: add TOOLCHAIN_WITH_SYSLINUX switch
Enable a choice between sys-freedos and syslinux as the boot loader
installer for floppy images.

In the case of syslinux, installation to a boot sector requires
root permissions regardless of whether the target is a device or an
image. The implementation therefore retains sys-freedos as the
toolchain's default installer unless TOOLCHAIN_WITH_SYSLINUX=y has
been explicitly specified.